Split project and cluster mgr into cluster or project based on use
