+++ pkg/controllers/management/auth/project_cluster_handler.go
@@ -3,21 +3,14 @@ package auth
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"reflect"
 	"time"
 
 	"github.com/rancher/norman/condition"
 	apisv3 "github.com/rancher/rancher/pkg/apis/management.cattle.io/v3"
-	"github.com/rancher/rancher/pkg/controllers"
-	"github.com/rancher/rancher/pkg/controllers/managementuserlegacy/systemimage"
-	wranglerv3 "github.com/rancher/rancher/pkg/generated/controllers/management.cattle.io/v3"
 	corev1 "github.com/rancher/rancher/pkg/generated/norman/core/v1"
 	v3 "github.com/rancher/rancher/pkg/generated/norman/management.cattle.io/v3"
-	rbacv1 "github.com/rancher/rancher/pkg/generated/norman/rbac.authorization.k8s.io/v1"
-	"github.com/rancher/rancher/pkg/project"
-	"github.com/rancher/rancher/pkg/settings"
 	"github.com/rancher/rancher/pkg/systemaccount"
 	"github.com/rancher/rancher/pkg/types/config"
 	"github.com/sirupsen/logrus"
@@ -25,7 +18,6 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
 )
 
@@ -35,135 +27,38 @@ const (
 	roleTemplatesRequired         = "authz.management.cattle.io/creator-role-bindings"
 )
 
-var defaultProjectLabels = labels.Set(map[string]string{"authz.management.cattle.io/default-project": "true"})
-var systemProjectLabels = labels.Set(map[string]string{"authz.management.cattle.io/system-project": "true"})
 var crtbCreatorOwnerAnnotations = map[string]string{creatorOwnerBindingAnnotation: "true"}
 
 func newPandCLifecycles(management *config.ManagementContext) (*projectLifecycle, *clusterLifecycle) {
 	m := &mgr{
-		mgmt:                 management,
-		nsLister:             management.Core.Namespaces("").Controller().Lister(),
-		prtbLister:           management.Management.ProjectRoleTemplateBindings("").Controller().Lister(),
-		crtbLister:           management.Management.ClusterRoleTemplateBindings("").Controller().Lister(),
-		crtbClient:           management.Management.ClusterRoleTemplateBindings(""),
-		projectLister:        management.Management.Projects("").Controller().Lister(),
-		projects:             management.Wrangler.Mgmt.Project(),
-		roleTemplateLister:   management.Management.RoleTemplates("").Controller().Lister(),
-		systemAccountManager: systemaccount.NewManager(management),
-		rbLister:             management.RBAC.RoleBindings("").Controller().Lister(),
-		roleBindings:         management.RBAC.RoleBindings(""),
+		mgmt:       management,
+		nsLister:   management.Core.Namespaces("").Controller().Lister(),
+		prtbLister: management.Management.ProjectRoleTemplateBindings("").Controller().Lister(),
+		crtbLister: management.Management.ClusterRoleTemplateBindings("").Controller().Lister(),
 	}
 	p := &projectLifecycle{
-		mgr: m,
+		mgr:                  m,
+		crtbClient:           management.Management.ClusterRoleTemplateBindings(""),
+		rbLister:             management.RBAC.RoleBindings("").Controller().Lister(),
+		roleBindings:         management.RBAC.RoleBindings(""),
+		systemAccountManager: systemaccount.NewManager(management),
 	}
 	c := &clusterLifecycle{
-		mgr: m,
+		mgr:                m,
+		projects:           management.Wrangler.Mgmt.Project(),
+		projectLister:      management.Management.Projects("").Controller().Lister(),
+		rbLister:           management.RBAC.RoleBindings("").Controller().Lister(),
+		roleBindings:       management.RBAC.RoleBindings(""),
+		roleTemplateLister: management.Management.RoleTemplates("").Controller().Lister(),
 	}
 	return p, c
 }
 
 type mgr struct {
-	mgmt          *config.ManagementContext
-	nsLister      corev1.NamespaceLister
-	projectLister v3.ProjectLister
-	projects      wranglerv3.ProjectClient
-
-	prtbLister           v3.ProjectRoleTemplateBindingLister
-	crtbLister           v3.ClusterRoleTemplateBindingLister
-	crtbClient           v3.ClusterRoleTemplateBindingInterface
-	roleTemplateLister   v3.RoleTemplateLister
-	systemAccountManager *systemaccount.Manager
-	rbLister             rbacv1.RoleBindingLister
-	roleBindings         rbacv1.RoleBindingInterface
-}
-
-func (m *mgr) createDefaultProject(obj runtime.Object) (runtime.Object, error) {
-	return m.createProject(project.Default, apisv3.ClusterConditionDefaultProjectCreated, obj, defaultProjectLabels)
-}
-
-func (m *mgr) createSystemProject(obj runtime.Object) (runtime.Object, error) {
-	return m.createProject(project.System, apisv3.ClusterConditionSystemProjectCreated, obj, systemProjectLabels)
-}
-
-func (m *mgr) createProject(name string, cond condition.Cond, obj runtime.Object, labels labels.Set) (runtime.Object, error) {
-	return cond.DoUntilTrue(obj, func() (runtime.Object, error) {
-		metaAccessor, err := meta.Accessor(obj)
-		if err != nil {
-			return obj, err
-		}
-		// Attempt to use the cache first
-		projects, err := m.projectLister.List(metaAccessor.GetName(), labels.AsSelector())
-		if err != nil || len(projects) > 0 {
-			return obj, err
-		}
-
-		// Cache failed, try the API
-		projects2, err := m.projects.List(metaAccessor.GetName(), v1.ListOptions{LabelSelector: labels.String()})
-		if err != nil || len(projects2.Items) > 0 {
-			return obj, err
-		}
-
-		annotation := map[string]string{}
-
-		creatorID := metaAccessor.GetAnnotations()[creatorIDAnn]
-		if creatorID != "" {
-			annotation[creatorIDAnn] = creatorID
-		}
-
-		if name == project.System {
-			latestSystemVersion, err := systemimage.GetSystemImageVersion()
-			if err != nil {
-				return obj, err
-			}
-			annotation[project.SystemImageVersionAnn] = latestSystemVersion
-		}
-
-		project := &apisv3.Project{
-			ObjectMeta: v1.ObjectMeta{
-				GenerateName: "p-",
-				Annotations:  annotation,
-				Labels:       labels,
-			},
-			Spec: apisv3.ProjectSpec{
-				DisplayName: name,
-				Description: fmt.Sprintf("%s project created for the cluster", name),
-				ClusterName: metaAccessor.GetName(),
-			},
-		}
-		updated, err := m.addRTAnnotation(project, "project")
-		if err != nil {
-			return obj, err
-		}
-		project = updated.(*apisv3.Project)
-		logrus.Infof("[%v] Creating %s project for cluster %v", clusterCreateController, name, metaAccessor.GetName())
-		if _, err = m.mgmt.Management.Projects(metaAccessor.GetName()).Create(project); err != nil {
-			return obj, err
-		}
-		return obj, nil
-	})
-}
-
-// deleteSystemProject deletes the system project(s) for a cluster in preparation for deleting the cluster namespace.
-// Normally, the webhook prevents deleting the system project, so Rancher needs to use the sudo user to force it.
-// Otherwise, the deleted namespace will be stuck terminating because it cannot garbage collect the project.
-func (m *mgr) deleteSystemProject(cluster *apisv3.Cluster, controller string) error {
-	bypassClient, err := m.projects.WithImpersonation(controllers.WebhookImpersonation())
-	if err != nil {
-		return fmt.Errorf("[%s] failed to create impersonation client: %w", controller, err)
-	}
-	projects, err := m.projectLister.List(cluster.Name, systemProjectLabels.AsSelector())
-	if err != nil {
-		return fmt.Errorf("[%s] failed to list projects: %w", controller, err)
-	}
-	var deleteError error
-	for _, p := range projects {
-		logrus.Infof("[%s] Deleting project %s", controller, p.Name)
-		err = bypassClient.Delete(p.Namespace, p.Name, nil)
-		if err != nil {
-			deleteError = errors.Join(deleteError, fmt.Errorf("[%s] failed to delete project '%s/%s': %w", controller, p.Namespace, p.Name, err))
-		}
-	}
-	return deleteError
+	mgmt       *config.ManagementContext
+	nsLister   corev1.NamespaceLister
+	prtbLister v3.ProjectRoleTemplateBindingLister
+	crtbLister v3.ClusterRoleTemplateBindingLister
 }
 
 func (m *mgr) reconcileCreatorRTB(obj runtime.Object) (runtime.Object, error) {
@@ -372,75 +267,6 @@ func (m *mgr) reconcileResourceToNamespace(obj runtime.Object, controller string
 
 		return obj, nil
 	})
-}
-
-func (m *mgr) addRTAnnotation(obj runtime.Object, context string) (runtime.Object, error) {
-	meta, err := meta.Accessor(obj)
-	if err != nil {
-		return obj, err
-	}
-
-	// If the annotation is already there move along
-	if _, ok := meta.GetAnnotations()[roleTemplatesRequired]; ok {
-		return obj, nil
-	}
-
-	rt, err := m.roleTemplateLister.List("", labels.NewSelector())
-	if err != nil {
-		return obj, err
-	}
-
-	annoMap := make(map[string][]string)
-
-	var restrictedAdmin bool
-	if settings.RestrictedDefaultAdmin.Get() == "true" {
-		restrictedAdmin = true
-	}
-
-	annoMap["created"] = []string{}
-	annoMap["required"] = []string{}
-
-	switch context {
-	case "project":
-		// If we are in restricted mode, ensure the default projects are not granting
-		// permissions to the restricted-admin
-		if restrictedAdmin {
-			proj := obj.(*apisv3.Project)
-			if proj.Spec.ClusterName == "local" && (proj.Spec.DisplayName == "Default" || proj.Spec.DisplayName == "System") {
-				break
-			}
-		}
-
-		for _, role := range rt {
-			if role.ProjectCreatorDefault && !role.Locked {
-				annoMap["required"] = append(annoMap["required"], role.Name)
-			}
-		}
-	case "cluster":
-		// If we are in restricted mode, ensure we don't give the default restricted-admin
-		// the default permissions in the cluster
-		if restrictedAdmin && meta.GetName() == "local" {
-			break
-		}
-
-		for _, role := range rt {
-			if role.ClusterCreatorDefault && !role.Locked {
-				annoMap["required"] = append(annoMap["required"], role.Name)
-			}
-		}
-	}
-
-	d, err := json.Marshal(annoMap)
-	if err != nil {
-		return obj, err
-	}
-
-	// Save the required role templates to the annotation on the obj
-	if meta.GetAnnotations() == nil {
-		meta.SetAnnotations(make(map[string]string))
-	}
-	meta.GetAnnotations()[roleTemplatesRequired] = string(d)
-	return obj, nil
 }
 
 func (m *mgr) updateClusterAnnotationandCondition(cluster *apisv3.Cluster, anno string, updateCondition bool) error {
